requirements.txt
scikit-learn
openpyxl
open3d
smplx
trimesh
imageio
spconv
easydict

gradio
litserve==0.2.3
pydantic==2.9.2
gdown
yacs
progress
ultralytics

supervision==0.23.0 ; python_full_version > '3.10'
transformers==4.44.2
uvloop==0.20.0

spectacularAI==1.35.0
progress==1.6.1

# lingbot-depth
xformers==0.0.28
0.0.27.post2
lightning 
imageio>=2.33.1
Pillow>=10.1.0
imageio-ffmpeg
einops
tqdm
ipdb
transformers
termcolor
# numpy==1.26.4
numpy==1.23.1
opencv-python 
scipy==1.16.1
matplotlib
h5py
tyro==0.9.2
open3d
gradio==4.44.1
gradio-imageslider==0.0.20
spaces
# 

# VGGT
huggingface_hub
einops
safetensors

scikit-image
progress 
ultralytics
mmcv==1.5.0
# torch_scatter
spconv
timm
einops
smplx
natsort
setuptools==57.0.0

hydra-core
accelerate
prettytable
